Correccion de que no guardaba el BarCode en Promociones costo envio

+++ Code/ServicesManagementCentral.Web/Controllers/PromocionesCostoEnvioController.cs
@@ -377,7 +377,7 @@ namespace ServicesManagement.Web.Controllers
         public ActionResult AddPromociones(int cnscPromo, int IdTipoCatalogo, int? IdOwner, int? IdFormatoTienda, string PostalCodeOrig, string PostalCodeDestino
             , string CiudadOrig, string CiudadDest, string EdoOrig, string EdoDest, int? IdSupplierWH, string SupplierName, int? IdTransportista, string NombreTransportista,
              int? IdTipoEnvio, int? IdTipoServicio, string Cve_CategSAP, string Desc_CategSAP, string Cve_GciaCategSAP, string Desc_GciaCategSAP
-            , string Material_MATNR, int? Id_Num_CodBarra, string nombre_SKU, decimal PesoMinimo, decimal PesoMaximo, int? IdTipoLogistica, int MesesSinIntereses
+            , string Material_MATNR, long? Id_Num_CodBarra, string nombre_SKU, decimal PesoMinimo, decimal PesoMaximo, int? IdTipoLogistica, int MesesSinIntereses
             , decimal ComprasMayor, decimal ComprasMenor, string FechaInicioPromo, string HoraInicioPromo, string FechaFinPromo, string HoraFinPromo
             , decimal? CostoEspecial, decimal? TarifaDesc)
         {
