+++ Code/ServicesManagementCentral.Web/DAL/DALPromocionesCostoEnvio.cs
@@ -223,7 +223,7 @@
         public static void PromocionesCostoEnvio_iUp(int IdTipoCatalogo, int? IdOwner, int? IdFormatoTienda, string PostalCodeOrig, string PostalCodeDestino
             , string CiudadOrig, string CiudadDest, string EdoOrig, string EdoDest, int? IdSupplierWH, string SupplierName, int? IdTransportista, string NombreTransportista
             , int? IdTipoEnvio, int? IdTipoServicio, string Cve_CategSAP, string Desc_CategSAP, string Cve_GciaCategSAP, string Desc_GciaCategSAP
-            , string Material_MATNR, int? Id_Num_CodBarra, string nombre_SKU, decimal? PesoMinimo, decimal? PesoMaximo, int? IdTipoLogistica, int? MesesSinIntereses
+            , string Material_MATNR, long? Id_Num_CodBarra, string nombre_SKU, decimal? PesoMinimo, decimal? PesoMaximo, int? IdTipoLogistica, int? MesesSinIntereses
             , decimal? ComprasMayor, decimal? ComprasMenor, DateTime FechaInicioPromo, TimeSpan HoraInicioPromo, DateTime FechaFinPromo, TimeSpan HoraFinPromo
             , decimal? CostoEspecial, decimal? TarifaDesc, DateTime FechaCreacion, TimeSpan HoraCreacion, string UsuarioCreacion, bool BitActivo)
 
@@ -297,7 +297,7 @@
         public static void PromocionesCostoEnvio_uUp(int cnscPromo, int IdTipoCatalogo, int? IdOwner, int? IdFormatoTienda, string PostalCodeOrig, string PostalCodeDestino
       , string CiudadOrig, string CiudadDest, string EdoOrig, string EdoDest, int? IdSupplierWH, string SupplierName, int? IdTransportista, string NombreTransportista
       , int? IdTipoEnvio, int? IdTipoServicio, string Cve_CategSAP, string Desc_CategSAP, string Cve_GciaCategSAP, string Desc_GciaCategSAP
-      , string Material_MATNR, int? Id_Num_CodBarra, string nombre_SKU, decimal? PesoMinimo, decimal? PesoMaximo, int? IdTipoLogistica, int? MesesSinIntereses
+      , string Material_MATNR, long? Id_Num_CodBarra, string nombre_SKU, decimal? PesoMinimo, decimal? PesoMaximo, int? IdTipoLogistica, int? MesesSinIntereses
       , decimal? ComprasMayor, decimal? ComprasMenor, DateTime FechaInicioPromo, TimeSpan HoraInicioPromo, DateTime FechaFinPromo, TimeSpan HoraFinPromo
       , decimal? CostoEspecial, decimal? TarifaDesc, DateTime FechaUltModif, TimeSpan HoraUltModif, string UsuarioUltModif, bool BitActivo)
